♻️ Extract option_type.go
